requirements.txt
watchdog
Pillow